First sanity check for Referentiel
