+++ api/notation/referentiel.py
@@ -5,6 +5,10 @@ from typing import Tuple, Dict, List
 from api.models.generated.action_referentiel import ActionReferentiel
 
 
+class ReferentielValueError(Exception):
+    pass
+
+
 class Referentiel:
     """Referentiel.
 
@@ -19,6 +23,7 @@ class Referentiel:
     """
 
     def __init__(self, root_action: ActionReferentiel):
+        self.__sanity_check(root_action)
         self.root_action: ActionReferentiel = root_action
         self.points: Dict[Tuple, float] = {}
         self.percentages: Dict[Tuple, float] = {}
@@ -26,7 +31,7 @@ class Referentiel:
         self.indices: List[Tuple]
         self.forward: List[Tuple]
         self.backward: List[Tuple]
-        self.__build_indices()
+        self.__build_indices_and_actions()
         self.__build_points()
         self.__build_percentages()
 
@@ -39,7 +44,7 @@ class Referentiel:
         parent = index[:-1]
         return self.children(parent)
 
-    def __build_indices(self):
+    def __build_indices_and_actions(self):
         """Build all indices lists"""
         self.indices: List[Tuple] = []
 
@@ -56,9 +61,7 @@ class Referentiel:
         add_action(self.root_action)
 
         self.forward: List[Tuple] = sorted(self.indices, key=lambda i: len(i))
-        self.backward: List[Tuple] = sorted(
-            self.indices, key=lambda i: len(i), reverse=True
-        )
+        self.backward: List[Tuple] = self.forward[::-1]
 
     def __build_points(self):
         """Build points
@@ -74,7 +77,7 @@ class Referentiel:
                 points = 500  # TODO: this should not be hard-coded
             elif len(index) == 1:
                 # axe
-                points = 100
+                points = 100  # TODO: this should not be hard-coded
             else:
                 # orientation, niveau, tache
                 points = max(self.actions[index].points, 0) * (
@@ -96,3 +99,10 @@ class Referentiel:
                         self.points[index] / self.points[index[:-1]]
                     )
         self.percentages[tuple()] = 1
+
+    @staticmethod
+    def __sanity_check(root_action: ActionReferentiel):
+        if root_action.id_nomenclature != "":
+            raise ReferentielValueError(
+                f"Root action should have '' as `id_nomenclature`, received {root_action.id_nomenclature} instead. "
+            )
