+++ tests/test_referentiel.py
@@ -1,33 +1,12 @@
-from typing import List
-
 import pytest
 
-from api.models.generated.action_referentiel import ActionReferentiel
-from api.notation.referentiel import Referentiel
+from api.notation.referentiel import Referentiel, ReferentielValueError
 
-
-def make_action_referentiel(
-    id: str = "default_id",
-    id_nomenclature: str = "",
-    nom: str = "default_nom",
-    description: str = "",
-    thematique_id: str = "default_thematique_id",
-    points: float = 42,
-    actions: List[ActionReferentiel] = [],
-):
-    return ActionReferentiel(
-        id=id,
-        id_nomenclature=id_nomenclature,
-        nom=nom,
-        description=description,
-        thematique_id=thematique_id,
-        points=points,
-        actions=actions,
-    )
+from .utils.factory import make_action_referentiel
 
 
 def test_referentiel_fails_when_wrong_root_action():
-    with pytest.raises():
+    with pytest.raises(ReferentielValueError):
         referentiel = Referentiel(
             root_action=make_action_referentiel(id_nomenclature="not_empty")
         )
